+++ packages/commonjs/test/fixtures/function/class-body-property-keys-are-special-keywords/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  description: 'preserve the class body property keys even if they are special keywords',
+};