+++ packages/replace/test/fixtures/form/ternary-operator/output.js
@@ -1,0 +1,3 @@
+/* eslint-disable */
+first ? second : third;
+console.log(first, second, third);