+++ packages/commonjs/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @rollup/plugin-commonjs ChangeLog
 
+## v26.0.1
+
+_2024-06-05_
+
+### Bugfixes
+
+- fix: correct import of glob (04a15b5)
+
+## v26.0.0
+
+_2024-06-05_
+
+### Breaking Changes
+
+- chore!: bump glob's version (#1695)
+
+## v25.0.8
+
+_2024-05-22_
+
+### Bugfixes
+
+- fix: preserve the class body property keys even if they are special keywords (#1688)
+
 ## v25.0.7
 
 _2023-10-15_