+++ packages/dynamic-import-vars/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rollup/plugin-dynamic-import-vars ChangeLog
 
+## v2.1.3
+
+_2024-09-22_
+
+### Bugfixes
+
+- fix: simplify regex (#1751)
+
+## v2.1.2
+
+_2023-11-28_
+
+### Bugfixes
+
+- fix: Allow a "no files found" error to be emitted as warning (#1625)
+
 ## v2.1.1
 
 _2023-11-25_